test: refactor SHMEM triton tests for NV and ROCm into shared backend suite
Consolidate NVSHMEM and rocSHMEM Triton tests into a shared SHMEM test module,
update CI to run the new shared shmem test

Future refactoring is needed to use shared layer _shmem_triton.py for NVSHMEM
backend to inherit common functionality for init_hook, require_nvshmem decorator
building and such.
