+++ test/distributed/test_shmem_triton.py
@@ -1,84 +1,37 @@
 # Owner(s): ["oncall: distributed"]
 
 import sys
+import unittest
 
 import torch
+import torch.distributed._symmetric_memory as symm_mem
 
-
-# Currently this common shmem_triton tests are tested on ROCm platforms only.
-# In a followup refactor test_nvshmem_triton.py can switch to this shmem_triton implementation.
-if torch.version.hip is None:
-    print("rocSHMEM Triton tests run on ROCm only; skipping on non-ROCm CI.")
+if not torch.backends.cuda.is_built() or not symm_mem.is_nvshmem_available():
+    print("SHMEM backend (NVSHMEM/rocSHMEM) not available, skipping tests")
     sys.exit(0)
 
-import unittest
+# Shared SHMEM Triton tests for both NVSHMEM (CUDA) and rocSHMEM (ROCm).
 
 import triton.language as tl
 
 import torch.distributed as dist
-import torch.distributed._symmetric_memory as symm_mem
 import torch.distributed._symmetric_memory._shmem_triton as shmem_triton
 from torch._inductor.runtime.triton_compat import triton
 from torch.testing._internal.common_distributed import MultiProcContinuousTest
 from torch.testing._internal.common_utils import (
     instantiate_parametrized_tests,
+    parametrize,
     run_tests,
+    skip_but_pass_in_sandcastle_if,
 )
 from torch.testing._internal.inductor_utils import IS_H100, requires_triton
-from torch.utils._triton import has_triton
 
 
 shmem_backend = shmem_triton.get_shmem_backend_module()
 requires_shmem = shmem_triton.requires_shmem
 
-
 device_type = "cuda"
 device_module = torch.get_device_module(device_type)
-
-
-class ShmemBackendMixin:
-    backend_name = "NVSHMEM"
-
-    def is_backend_available(self) -> bool:
-        return symm_mem.is_nvshmem_available()
-
-    def skip_or_xfail_reason(self, op: str) -> str | None:
-        return None
-
-    def requires_h100(self) -> bool:
-        return False
-
-
-class NVShmemBackendMixin(ShmemBackendMixin):
-    backend_name = "NVSHMEM"
-
-    def requires_h100(self) -> bool:
-        return True
-
-    def skip_or_xfail_reason(self, op: str) -> str | None:
-        if self.requires_h100() and not IS_H100:
-            return "NVSHMEM Triton tests require H100."
-        return None
-
-
-class RocShmemBackendMixin(ShmemBackendMixin):
-    # Same backend id as CUDA NVSHMEM path; implementation is rocSHMEM on HIP.
-    backend_name = "NVSHMEM"
-
-    def skip_or_xfail_reason(self, op: str) -> str | None:
-        if torch.version.hip is None:
-            return "rocSHMEM Triton tests only run on ROCm."
-        if op in {
-            "test_triton_alltoall",
-            "test_triton_broadcast",
-            "test_triton_sum_reduce",
-            "test_triton_minmax_reduce",
-            "test_triton_prod_reduce",
-        }:
-            return "rocSHMEM *_wg collective symbols are unavailable in current device bitcode for this op."
-        if op == "test_triton_put_signal_add":
-            return "Known hang in rocSHMEM Triton put_signal_add path."
-        return None
 
 
 @requires_shmem
@@ -121,8 +74,25 @@ def my_signal_wait_until_kernel(signal, cmp_op, cmp_val):
 
 @requires_shmem
 @triton.jit
+def my_signal_op_kernel(
+    sig_addr,
+    signal,
+    sig_op,
+    peer,
+):
+    shmem_backend.signal_op(sig_addr, signal, sig_op, peer)
+
+
+@requires_shmem
+@triton.jit
 def my_wait_until_kernel(ivar, cmp_op, cmp_val):
     shmem_backend.wait_until(ivar, cmp_op, cmp_val)
+
+
+@requires_shmem
+@triton.jit
+def my_fence_kernel():
+    shmem_backend.fence()
 
 
 @requires_shmem
@@ -159,7 +129,103 @@ def my_put_with_quiet_kernel(
     shmem_backend.put(flag_dst, flag_src, 1, peer)
 
 
+@requires_shmem
+@triton.jit
+def my_barrier_test_kernel(dst, src, nelems):
+    # Validate device-side barrier by coordinating put/compute phases in one launch.
+    my_pe = shmem_backend.my_pe()
+    n_pes = shmem_backend.n_pes()
+
+    if my_pe == 0:
+        p_src = src.to(tl.pointer_type(tl.int32))
+        tl.store(p_src, 42)
+        i = 1
+        while i < n_pes:
+            shmem_backend.put(dst, src, nelems, i)
+            i += 1
+
+    shmem_backend.barrier_all()
+
+    if my_pe != 0:
+        p_dst = dst.to(tl.pointer_type(tl.int32))
+        received = tl.load(p_dst)
+        tl.store(p_dst, received + 1)
+
+
+@requires_shmem
+@triton.jit
+def my_sync_test_kernel(local_data, remote_data, nelems):
+    my_pe = shmem_backend.my_pe()
+    n_pes = shmem_backend.n_pes()
+
+    p_local = local_data.to(tl.pointer_type(tl.int32))
+    unique_value = my_pe + 100
+    tl.store(p_local, unique_value)
+
+    # sync_all makes local writes visible before subsequent remote reads.
+    shmem_backend.sync_all()
+
+    next_pe = (my_pe + 1) % n_pes
+    shmem_backend.get(remote_data, local_data, nelems, next_pe)
+
+
+@requires_shmem
+@triton.jit
+def my_barrier_all_kernel():
+    shmem_backend.barrier_all()
+
+
+@requires_shmem
+@triton.jit
+def my_alltoall_kernel(
+    team_handle,
+    dst,
+    src,
+    nelems_per_pe,
+):
+    shmem_backend.alltoall(team_handle, dst, src, nelems_per_pe)
+
+
+@requires_shmem
+@triton.jit
+def my_broadcast_kernel(
+    team_handle,
+    dst,
+    src,
+    nelems,
+    pe_root,
+):
+    shmem_backend.broadcast(team_handle, dst, src, nelems, pe_root)
+
+
+@requires_shmem
+@triton.jit
+def my_reduce_kernel(
+    team_handle,
+    dest_tensor,
+    source_tensor,
+    nreduce,
+    operation: tl.constexpr,
+):
+    shmem_backend.reduce(team_handle, dest_tensor, source_tensor, nreduce, operation)
+
+
 class ShmemTritonTestBase(MultiProcContinuousTest):
+    # """
+    # Abstract base class for SHMEM Triton tests.
+    #
+    # This class provides a unified base for SHMEM tests (NVSHMEM, rocSHMEM).
+    # Backend-specific skip policy is expressed with explicit decorators.
+    #
+    # SHMEMTritonTest is the single concrete test class that is collected.
+    #
+    # For backend-specific tests:
+    # - If few, add methods to SHMEMTritonTestBase with:
+    #     @unittest.skipIf(torch.version.hip is not None, "NVSHMEM-only")
+    #     @unittest.skipIf(torch.version.hip is None, "rocSHMEM-only")
+    # - If many, split into NVSHMEMTritonTest/ROCSHMEMTritonTest subclasses with class-level skips.
+    # Start with decorated methods here; refactor if backend-only tests grow.
+    __test__ = False
     backend_name = "NVSHMEM"
 
     def setUp(self) -> None:
@@ -171,22 +237,12 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
     def device(self) -> torch.device:
         return torch.device(device_type, self.rank)
 
-    def _skip_unavailable(self, op_name: str) -> None:
-        if not has_triton():
-            self.skipTest("Triton not available")
-        if not self.is_backend_available():
-            self.skipTest(f"{self.backend_name}/rocSHMEM backend not available")
-        reason = self.skip_or_xfail_reason(op_name)
-        if reason:
-            self.skipTest(reason)
-
     def _init_device(self) -> None:
         device_module.set_device(self.device)
         symm_mem.set_backend(self.backend_name)
 
     @requires_triton()
     def test_triton_put(self) -> None:
-        self._skip_unavailable("test_triton_put")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -217,7 +273,6 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
             )
 
     def _run_triton_get(self, nbi: bool) -> None:
-        self._skip_unavailable("test_triton_get")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -245,7 +300,6 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
             )
 
     def _run_triton_get_ring(self) -> None:
-        self._skip_unavailable("test_triton_get_ring")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -270,12 +324,9 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
         )
 
     @requires_triton()
-    def test_triton_get(self) -> None:
-        self._run_triton_get(nbi=False)
-
-    @requires_triton()
-    def test_triton_get_nbi(self) -> None:
-        self._run_triton_get(nbi=True)
+    @parametrize("nbi", [False, True])
+    def test_triton_get(self, nbi: bool) -> None:
+        self._run_triton_get(nbi=nbi)
 
     @requires_triton()
     def test_triton_get_ring(self) -> None:
@@ -283,7 +334,6 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
 
     @requires_triton()
     def test_triton_wait_until(self) -> None:
-        self._skip_unavailable("test_triton_wait_until")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -315,7 +365,6 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
 
     @requires_triton()
     def test_triton_signal_wait_until(self) -> None:
-        self._skip_unavailable("test_triton_signal_wait_until")
         self._init_device()
 
         group_name = dist.distributed_c10d._get_default_group().group_name
@@ -364,7 +413,6 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
 
     @requires_triton()
     def test_triton_fence(self) -> None:
-        self._skip_unavailable("test_triton_fence")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -417,7 +465,6 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
 
     @requires_triton()
     def test_triton_quiet(self) -> None:
-        self._skip_unavailable("test_triton_quiet")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -459,8 +506,74 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
         dist.barrier()
 
     @requires_triton()
+    def test_triton_barrier(self) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        rank = self.rank
+        numel = 1
+        dtype = torch.int32
+
+        src = symm_mem.empty(numel, dtype=dtype, device=self.device).fill_(0)
+        dst = symm_mem.empty(numel, dtype=dtype, device=self.device).fill_(0)
+        symm_mem.rendezvous(src, group=group_name)
+        symm_mem.rendezvous(dst, group=group_name)
+
+        my_barrier_test_kernel[(1,)](
+            dst,
+            src,
+            nelems=numel,
+            launch_cooperative_grid=True,
+            num_ctas=1,
+        )
+        dist.barrier()
+
+        if rank == 0:
+            torch.testing.assert_close(
+                src, torch.tensor([42], device=self.device, dtype=dtype)
+            )
+        else:
+            torch.testing.assert_close(
+                dst, torch.tensor([43], device=self.device, dtype=dtype)
+            )
+
+    @requires_triton()
+    def test_triton_sync(self) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        rank = self.rank
+        numel = 1
+        dtype = torch.int32
+
+        local_data = symm_mem.empty(numel, dtype=dtype, device=self.device).fill_(0)
+        remote_data = symm_mem.empty(numel, dtype=dtype, device=self.device).fill_(0)
+        symm_mem.rendezvous(local_data, group=group_name)
+        symm_mem.rendezvous(remote_data, group=group_name)
+
+        my_sync_test_kernel[(1,)](
+            local_data,
+            remote_data,
+            nelems=numel,
+            launch_cooperative_grid=True,
+            num_ctas=1,
+        )
+
+        expected_local = rank + 100
+        torch.testing.assert_close(
+            local_data, torch.tensor([expected_local], device=self.device, dtype=dtype)
+        )
+
+        next_rank = (rank + 1) % self.world_size
+        expected_remote = next_rank + 100
+        torch.testing.assert_close(
+            remote_data, torch.tensor([expected_remote], device=self.device, dtype=dtype)
+        )
+
+    @requires_triton()
     def test_triton_put_signal_set(self) -> None:
-        self._skip_unavailable("test_triton_put_signal_set")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -508,8 +621,11 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
             )
 
     @requires_triton()
+    @unittest.skipIf(
+        torch.version.hip is not None,
+        "Known hang in rocSHMEM Triton put_signal_add path.",
+    )
     def test_triton_put_signal_add(self) -> None:
-        self._skip_unavailable("test_triton_put_signal_add")
         torch.manual_seed(42 + self.rank)
         self._init_device()
 
@@ -556,47 +672,273 @@ class ShmemTritonTestBase(MultiProcContinuousTest):
                 flag, torch.tensor([SIGNAL_VAL], dtype=torch.int64, device=self.device)
             )
 
-    @unittest.skip("Collective launch stays backend-specific for now.")
     @requires_triton()
-    def test_triton_alltoall(self) -> None:
-        pass
-
-    @unittest.skip("Collective launch stays backend-specific for now.")
-    @requires_triton()
-    def test_triton_broadcast(self) -> None:
-        pass
-
-    @unittest.skip("Collective launch stays backend-specific for now.")
-    @requires_triton()
-    def test_triton_sum_reduce(self) -> None:
-        pass
-
-    @unittest.skip("Collective launch stays backend-specific for now.")
-    @requires_triton()
-    def test_triton_minmax_reduce(self) -> None:
-        pass
-
-    @unittest.skip("Collective launch stays backend-specific for now.")
-    @requires_triton()
-    def test_triton_prod_reduce(self) -> None:
-        pass
-
-
-if torch.version.hip is not None:
-    ActiveShmemBackendMixin = RocShmemBackendMixin
-else:
-    ActiveShmemBackendMixin = NVShmemBackendMixin
-
-
-@instantiate_parametrized_tests
-class SHMEMTritonTest(ActiveShmemBackendMixin, ShmemTritonTestBase):
     @unittest.skipIf(
         torch.version.hip is not None,
-        "Known hang in rocSHMEM Triton put_signal_add path.",
+        "rocSHMEM *_wg collective symbols are unavailable in current device bitcode for this op.",
     )
+    def test_triton_alltoall(self) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        world_size = dist.get_world_size()
+        rank = self.rank
+        nelems_per_pe = 2
+        dtype = torch.int64
+        src_size = nelems_per_pe * world_size
+        src = symm_mem.empty(src_size, dtype=dtype, device=self.device)
+        for i in range(world_size):
+            value = rank * 100 + i
+            src[i * nelems_per_pe : (i + 1) * nelems_per_pe] = value
+        dst = symm_mem.empty(src_size, dtype=dtype, device=self.device).fill_(-1)
+        symm_mem.rendezvous(src, group=group_name)
+        symm_mem.rendezvous(dst, group=group_name)
+        dist.barrier()
+        team_handle = 0
+        my_alltoall_kernel[(1,)](
+            team_handle,
+            dst,
+            src,
+            nelems_per_pe,
+            launch_cooperative_grid=True,
+        )
+        dist.barrier()
+        for i in range(world_size):
+            expected = i * 100 + rank
+            actual = dst[i * nelems_per_pe : (i + 1) * nelems_per_pe]
+            torch.testing.assert_close(actual, torch.full_like(actual, expected))
+
     @requires_triton()
-    def test_triton_put_signal_add(self) -> None:
-        super().test_triton_put_signal_add()
+    @unittest.skipIf(
+        torch.version.hip is not None,
+        "rocSHMEM *_wg collective symbols are unavailable in current device bitcode for this op.",
+    )
+    def test_triton_broadcast(self) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        rank = self.rank
+        nelems = 4
+        dtype = torch.int64
+        pe_root = 0
+        src = symm_mem.empty(nelems, dtype=dtype, device=self.device)
+        dst = symm_mem.empty(nelems, dtype=dtype, device=self.device).fill_(-999)
+        if rank == pe_root:
+            for i in range(nelems):
+                src[i] = 100 + i
+        else:
+            src.fill_(-1)
+        symm_mem.rendezvous(src, group=group_name)
+        symm_mem.rendezvous(dst, group=group_name)
+        dist.barrier()
+        team_handle = 0
+        my_broadcast_kernel[(1,)](
+            team_handle,
+            dst,
+            src,
+            nelems,
+            pe_root,
+            launch_cooperative_grid=True,
+        )
+        dist.barrier()
+        expected = [100 + i for i in range(nelems)]
+        torch.testing.assert_close(
+            dst, torch.tensor(expected, device=self.device, dtype=dtype)
+        )
+
+    @requires_triton()
+    @unittest.skipIf(
+        torch.version.hip is not None,
+        "rocSHMEM *_wg collective symbols are unavailable in current device bitcode for this op.",
+    )
+    @parametrize(
+        "dtype",
+        [
+            torch.int8,
+            torch.int16,
+            torch.int32,
+            torch.int64,
+            torch.uint8,
+            torch.float16,
+            torch.float32,
+            torch.bfloat16,
+        ],
+    )
+    def test_triton_sum_reduce(self, dtype) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        world_size = dist.get_world_size()
+        nreduce = 3
+        src = symm_mem.empty(nreduce, dtype=dtype, device=self.device)
+        for i in range(nreduce):
+            src[i] = (self.rank + 1) * (i + 1)
+        dst = symm_mem.empty(nreduce, dtype=dtype, device=self.device).fill_(-1)
+        symm_mem.rendezvous(src, group=group_name)
+        symm_mem.rendezvous(dst, group=group_name)
+        expected = []
+        for i in range(nreduce):
+            total = sum((r + 1) * (i + 1) for r in range(world_size))
+            expected.append(total)
+        dist.barrier()
+        team_handle = 0
+        my_reduce_kernel[(1,)](
+            team_handle,
+            dst,
+            src,
+            nreduce,
+            operation="sum",
+            launch_cooperative_grid=True,
+        )
+        dist.barrier()
+        torch.testing.assert_close(
+            dst, torch.tensor(expected, device=self.device, dtype=dtype)
+        )
+
+    @requires_triton()
+    @unittest.skipIf(
+        torch.version.hip is not None,
+        "rocSHMEM *_wg collective symbols are unavailable in current device bitcode for this op.",
+    )
+    @parametrize(
+        "dtype",
+        [
+            torch.int8,
+            torch.int16,
+            torch.int32,
+            torch.int64,
+            torch.float16,
+            torch.float32,
+            torch.float64,
+            torch.bfloat16,
+        ],
+    )
+    def test_triton_minmax_reduce(self, dtype) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        world_size = dist.get_world_size()
+        nreduce = 2
+        src_min = symm_mem.empty(nreduce, dtype=dtype, device=self.device)
+        src_max = symm_mem.empty(nreduce, dtype=dtype, device=self.device)
+        for i in range(nreduce):
+            if i == 0:
+                src_min[i] = 10 + self.rank * 5
+                src_max[i] = 10 + self.rank * 5
+            else:
+                src_min[i] = 20 - self.rank * 15
+                src_max[i] = 20 - self.rank * 15
+        dst_min = symm_mem.empty(nreduce, dtype=dtype, device=self.device).fill_(-1)
+        dst_max = symm_mem.empty(nreduce, dtype=dtype, device=self.device).fill_(-1)
+        symm_mem.rendezvous(src_min, group=group_name)
+        symm_mem.rendezvous(src_max, group=group_name)
+        symm_mem.rendezvous(dst_min, group=group_name)
+        symm_mem.rendezvous(dst_max, group=group_name)
+        all_values = []
+        for i in range(nreduce):
+            values = []
+            for r in range(world_size):
+                if i == 0:
+                    values.append(10 + r * 5)
+                else:
+                    values.append(20 - r * 15)
+            all_values.append(values)
+        expected_min = [min(vals) for vals in all_values]
+        expected_max = [max(vals) for vals in all_values]
+        dist.barrier()
+        team_handle = 0
+        my_reduce_kernel[(1,)](
+            team_handle,
+            dst_min,
+            src_min,
+            nreduce,
+            operation="min",
+            launch_cooperative_grid=True,
+        )
+        my_reduce_kernel[(1,)](
+            team_handle,
+            dst_max,
+            src_max,
+            nreduce,
+            operation="max",
+            launch_cooperative_grid=True,
+        )
+        dist.barrier()
+        torch.testing.assert_close(
+            dst_min, torch.tensor(expected_min, device=self.device, dtype=dtype)
+        )
+        torch.testing.assert_close(
+            dst_max, torch.tensor(expected_max, device=self.device, dtype=dtype)
+        )
+
+    @requires_triton()
+    @unittest.skipIf(
+        torch.version.hip is not None,
+        "rocSHMEM *_wg collective symbols are unavailable in current device bitcode for this op.",
+    )
+    @parametrize(
+        "dtype",
+        [
+            torch.int8,
+            torch.int16,
+            torch.int32,
+            torch.int64,
+            torch.float16,
+            torch.float32,
+            torch.bfloat16,
+        ],
+    )
+    def test_triton_prod_reduce(self, dtype) -> None:
+        torch.manual_seed(42 + self.rank)
+        self._init_device()
+        group_name = dist.distributed_c10d._get_default_group().group_name
+        world_size = dist.get_world_size()
+        nreduce = 3
+        src = symm_mem.empty(nreduce, dtype=dtype, device=self.device)
+        for i in range(nreduce):
+            if i == 0:
+                src[i] = 1 if self.rank % 2 == 0 else 2
+            elif i == 1:
+                src[i] = 1
+            else:
+                src[i] = 1 if (self.rank // 2) % 2 == 0 else 2
+        dst = symm_mem.empty(nreduce, dtype=dtype, device=self.device).fill_(-1)
+        symm_mem.rendezvous(src, group=group_name)
+        symm_mem.rendezvous(dst, group=group_name)
+        vals = torch.empty(nreduce, world_size, dtype=dtype)
+        vals[0, ::2] = 1
+        vals[0, 1::2] = 2
+        vals[1] = 1
+        for rank in range(world_size):
+            vals[2, rank] = 1 if (rank // 2) % 2 == 0 else 2
+        expected = vals.prod(-1).tolist()
+        dist.barrier()
+        team_handle = 0
+        my_reduce_kernel[(1,)](
+            team_handle,
+            dst,
+            src,
+            nreduce,
+            operation="prod",
+            launch_cooperative_grid=True,
+        )
+        dist.barrier()
+        torch.testing.assert_close(
+            dst, torch.tensor(expected, device=self.device, dtype=dtype)
+        )
+
+
+instantiate_parametrized_tests(ShmemTritonTestBase)
+
+
+class SHMEMTritonTest(ShmemTritonTestBase):
+    __test__ = True
+
+# class-level skip for NVSHMEM Triton tests on non-H100 platforms.
+SHMEMTritonTest = skip_but_pass_in_sandcastle_if(
+    torch.version.hip is None and not IS_H100,
+    "NVSHMEM Triton tests require H100.",
+)(SHMEMTritonTest)
 
 
 if __name__ == "__main__":
